Only cache based on the URL for now
This is kind of a hack because we just happen to know that there is only
a single POST request, so there is no overlapping URLs right now.

Optimally we would cache based on the request content too, but
MemoryCache is kind of limited concerning comparisons and I am not
bothering to implement that right now.