Split out changelog into separate files
Since Jellyfin shows the entire changelog for all available versions on
the plugin page, it doesn't really make sense to keep expanding a
running changelog.

Instead of manually editing the build.yaml Jellyfin metadata file, it is
now generated at build time, from a template and the latest changelog
for the dotnet project version.

This also allows for the opportunity to reduce the number of places that
the version has to updated at to only a single occurrence, which will
hopefully be more robust.