Migrate to CUIDs for identifying items
The main reason for this is that the part numbers turned out not to be
stable across releases, so instead we use the CUIDs supplied by the API.
Those are also not stable, but are at least unique to a particular
deployment, so they properly invalidate when something changes instead
of misidentifying an arc or episode like the part-based IDs did.

This change also adds filename CRC-32 matching, because the API exposes
those as well now.